+++ src/soyboy/types.rs
@@ -51,6 +51,6 @@ pub trait AudioProcessor<T> {
 }
 
 pub trait Oscillator {
-    fn set_pitch(&mut self, midi_note_number: i16);
-    fn set_velocity(&mut self, velocity: f32);
+    fn set_pitch(&mut self, midi_note_number: u16);
+    fn set_velocity(&mut self, velocity: f64);
 }
