+++ src/soyboy/utils.rs
@@ -23,8 +23,8 @@ pub fn pulse(phase: f64, duty: f64) -> i4 {
 const NOTE_NUMBER_OF_440_HZ: i16 = 69;
 
 /// https://steinbergmedia.github.io/vst3_doc/vstinterfaces/structSteinberg_1_1Vst_1_1NoteOnEvent.html の pitch の項目
-pub fn frequency_from_note_number(note_num: i16) -> f64 {
-    440.0 * 2.0_f64.powf((note_num - NOTE_NUMBER_OF_440_HZ) as f64 / 12.0)
+pub fn frequency_from_note_number(note_num: u16) -> f64 {
+    440.0 * 2.0_f64.powf((note_num as i16 - NOTE_NUMBER_OF_440_HZ) as f64 / 12.0)
 }
 
 pub fn level(decibel: f64) -> f64 {
